Removing TxEnum. Fixing DatasetManager.
