+++ core/blockchain/blockchain_utils.py
@@ -3,8 +3,7 @@ import logging
 import requests
 import time
 from typing import Callable
-
-from core.blockchain.tx_utils import TxEnum, Transaction
+from enum import Enum
 
 
 logging.basicConfig(level=logging.DEBUG,
@@ -147,3 +146,30 @@ def setter(client: object, key: str, port: int, value: object,
     except Exception as e:
         logging.info("HTTP POST error, got: {0}".format(e))
     return tx_receipt.text
+
+##############################################################################
+###                              OBJECTS                                   ###
+##############################################################################
+
+class TxEnum(Enum):
+    """
+    Enum of tx-related constants, e.g. standard keys present in a transaction
+    """
+    KEY = "KEY"
+    CONTENT = "CONTENT"
+    MESSAGES = "MESSAGES"
+
+class Transaction(object):
+    """
+    Object that represents transactions with the specification outlined:
+        `{
+            'key': ...
+            'value': ...
+         }`
+    """
+    def __init__(self, key: str, value: str) -> None:
+        self.key = key
+        self.value = value
+
+    def get_tx(self) -> dict:
+       return {TxEnum.KEY.name: self.key, TxEnum.CONTENT.name: self.value}
