+++ core/dataset_manager.py
@@ -7,10 +7,11 @@ import os
 
 import numpy as np
 import pandas as pd
+import ipfsapi
 
 import core.utils.context
 from core.configuration import ConfigurationManager
-
+from core.blockchain.blockchain_utils import setter
 
 logging.basicConfig(level=logging.INFO,
                 format='[DatasetManager] %(asctime)s %(levelname)s %(message)s')
@@ -83,6 +84,11 @@ class DatasetManager():
         self.rfp = raw_filepath
         self.tfp = None
         self._validate_data()
+        self.host = config.get("BLOCKCHAIN", "host")
+        self.ipfs_port = config.getint("BLOCKCHAIN", "ipfs_port")
+        self.port = config.getint("BLOCKCHAIN", "http_port")
+        self.timeout = config.getint("BLOCKCHAIN", "timeout")
+        self.client = ipfsapi.connect(self.host, self.ipfs_port)
 
     def _validate_data(self):
         """
@@ -222,7 +228,7 @@ class DatasetManager():
             assert not os.path.isdir(self.tfp)
         self.tfp = None
 
-    def check_key_length(key):
+    def check_key_length(self, key):
         """
         Keys for datasets can only be at most 30 characters long.
         """
@@ -260,7 +266,7 @@ class DatasetManager():
             if 'md' not in folder_dict:
                 raise NoMetadataFoundError(folder)
             value[folder] = folder_dict
-        self.client.setter(name, value)
+        receipt = setter(client=self.client, key=name, value=value, port=self.port)
 
     def post_dataset(self, name):
         """
@@ -287,4 +293,4 @@ class DatasetManager():
             folder_dict['ds'] = sample.to_json()
             folder_dict['md'] = md.to_json()
             value[folder] = folder_dict
-        self.client.setter(name, value)
+        receipt = setter(client=self.client, key=name, value=value, port=self.port)
