+++ tests/artifacts/dataset_manager/configuration.ini
@@ -7,3 +7,9 @@ frequency_in_mins = 1
 
 [RUNNER]
 weights_directory = weights
+
+[BLOCKCHAIN]
+ipfs_port = 5001
+http_port = 3000
+host = 127.0.0.1
+timeout = 5
