+++ tests/test_blockchain_gateway.py
@@ -4,8 +4,7 @@ import pytest
 from core.configuration                 import ConfigurationManager	
 from core.blockchain.blockchain_gateway import BlockchainGateway
 from core.utils.enums                   import RawEventTypes
-from core.blockchain.tx_utils           import TxEnum
-from core.blockchain.blockchain_utils   import setter
+from core.blockchain.blockchain_utils   import setter, TxEnum
 
 @pytest.fixture	
 def config_manager():
