+++ tests/test_communication_manager.py
@@ -10,7 +10,7 @@ from core.configuration         import ConfigurationManager
 from tests.testing_utils        import make_initialize_job, make_model_json
 from tests.testing_utils        import make_serialized_job, serialize_job
 from core.utils.enums           import RawEventTypes, JobTypes, MessageEventTypes
-from core.utils.keras import serialize_weights
+from core.utils.keras           import serialize_weights
 
 
 config_manager = ConfigurationManager()
